refactor listener actions
